backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
email-validator>=2.0.0
dnspython>=2.0.0
python-dotenv
mangum
passlib
bcrypt==4.0.1
python-jose[cryptography]